refactor engine, modifier and FilesIO
