+++ src/index.ts
@@ -1,22 +1,35 @@
-// import * as PIXI from 'pixi.js';
+import * as PIXI from 'pixi.js';
 import './index.css';
 import App from './interfaces/App';
 import { PixiEngine } from './interfaces/PixiEngine';
 import { urlSmallSet as imageSet } from './fixtures/imagesDataSet';
 import { SnapShotContainer } from './interfaces/SnapShot';
-import { SpaceModifiers } from './interfaces/SpaceModifiers';
+import FilesIO from './interfaces/FilesIO';
+import { SpaceModifiers } from './modifiers/SpaceModifiers';
 
 async function main() {
   const appDiv = document.querySelector('.app');
   if (appDiv instanceof HTMLElement) {
     const app = new App(new PixiEngine(appDiv, 1000, 800));
 
-    const loader = await PixiEngine.loadUrlSet(imageSet);
+    /** Temp Test sample **/
+    let rectangle = new PIXI.Graphics();
+    rectangle.lineStyle(1.1, 0xff3300, 1);
+    rectangle.drawRect(300, 300, 100, 100);
+    app.engine.addToViewport(rectangle);
 
+    /** Load test images **/
+    const loader = await FilesIO.loadUrlSet(imageSet);
+
+    /** Begin store? **/
     app.state.snapshots.store = SnapShotContainer.createSnapShotsFromPixiResources(
       loader.resources,
     );
 
+    console.log('app.state.snapshots.store', app.state.snapshots.store);
+    app.engine.addToViewport(app.state.snapshots.store);
+
+    /** Auto update from store **/
     /** TODO: change modifier interface to matrix output (prev, next states for animation) **/
     SpaceModifiers.transformPositionGrid(
       app.state.snapshots.store,
@@ -26,13 +39,9 @@ async function main() {
       0.1,
     );
 
-    // Objects viewport initialization
-    app.state.snapshots.store.forEach((snapShot, i, store) => {
-      app.engine.pixiViewport.addChild(snapShot);
-    });
-
+    // /** Auto update from store **/
     // app.state.snapshots.store.forEach((snapShot, i) => {
-    // snapShot.x = 0;
+    //   snapShot.x = 0;
     // });
 
     // viewport.on('zoomed', (e) => {
