+++ src/interfaces/PixiEngine.ts
@@ -18,7 +18,9 @@ export class PixiEngine {
       resolution: 1,
       transparent: true,
     });
+    targetDiv.appendChild(this.pixiApp.view);
 
+    // Viewport
     this.pixiViewport = new Viewport({
       screenWidth: window.innerWidth,
       screenHeight: window.innerHeight,
@@ -29,35 +31,15 @@ export class PixiEngine {
 
     this.pixiViewport.drag().pinch().wheel().decelerate();
 
+    // Initial render of blank viewport
     this.pixiApp.stage.addChild(this.pixiViewport);
-
-    targetDiv.appendChild(this.pixiApp.view);
-
-    /** Temp Test sample **/
-    let rectangle = new PIXI.Graphics();
-    rectangle.lineStyle(1.1, 0xff3300, 1);
-    rectangle.drawRect(300, 300, 100, 100);
-    this.pixiViewport.addChild(rectangle);
   }
 
-  static loadUrlSet(urlSet: string[]): Promise<PIXI.Loader> {
-    return new Promise((resolve) => {
-      const loader = new PIXI.Loader();
-      loader
-        .add(urlSet)
-        .on('progress', PixiEngine.urlSetLoaderProgress)
-        .load((loader: PIXI.Loader) => resolve(loader));
-    });
-  }
+  addToViewport(displayObject: PIXI.DisplayObject | PIXI.DisplayObject[]) {
+    if (Array.isArray(displayObject)) {
+      return this.pixiViewport.addChild(...displayObject);
+    }
 
-  static urlSetLoaderProgress(
-    loader: PIXI.Loader,
-    resource: PIXI.LoaderResource,
-  ): void {
-    console.log('loading: ' + resource.url);
-    // console.log("progress: " + loader.progress + "%");
-    //If you gave your files names as the first argument
-    //of the `add` method, you can access them like this
-    //console.log("loading: " + resource.name);
+    return this.pixiViewport.addChild(displayObject);
   }
 }
